we got fn sponza

+++ src/engine/task_list.cpp
@@ -8,7 +8,7 @@ void add_gfx_task( TaskList& task_list, GfxTask task ) { task_list.gfx_tasks.pus
 
 void add_pipeline_barrier( TaskList& task_list, PipelineBarrierDescriptor barrier )
 {
-    task_list.pipeline_barriers.push_back( std::pair( task_list.gfx_tasks.size(), barrier ) );
+    task_list.pipeline_barriers.push_back( std::pair( static_cast<int>(task_list.gfx_tasks.size()), barrier ) );
 }
 
 } // namespace racecar::engine
